+++ src/index.html
@@ -45,7 +45,7 @@
             </nav>
           
             <form class="search__form">
-              <input type="text" class="search__text" placeholder="Поиск фильмов">
+              <input type="text" name="searchQuery" class="search__text" placeholder="Поиск фильмов">
               <button class="search__btn" type="submit">
                 <svg class="search__logo">
                   <use href="./images/icons.svg#icon-search"></use>
