--- src/js/temporaryInfiniteScroll.js
@@ -1,17 +1,0 @@
-import  {refs}  from './getRefs';
-
-function loadMore() {
-  const onEntry = entries => {
-    entries.forEach(entry => {
-      if (entry.isIntersecting && apiService.query !== '') {
-        getNewPage();
-      }
-    });
-  };
-
-  const options = {
-    rootMargin: '330px',
-  };
-  const observer = new IntersectionObserver(onEntry, options);
-  observer.observe(refs.observerElement);
-}
